Align the debug output with what libxkbcommon does
The documentation of libxkbcommon says: "Any Unicode/ISO 10646 character in the range U+0100 to U+10FFFF can be represented by a keysym value in the range 0x01000100 to 0x0110FFFF. The name of Unicode keysyms is U<codepoint>, e.g. UA1B2. The name of other unnamed keysyms is the hexadecimal representation of their value, e.g. 0xabcd1234"